+++ examples/issues.rs
@@ -2,7 +2,7 @@ extern crate hyper;
 extern crate hubcaps;
 
 use hyper::Client;
-use hubcaps::{Github, IssueListOptions};
+use hubcaps::{Credentials, Github, IssueListOptions};
 use std::env;
 
 fn main() {
@@ -11,7 +11,7 @@ fn main() {
             let client = Client::new();
             let github = Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
             &client,
-                                     Some(token));
+                                     Credentials::Token(token));
             let repo = github.repo("softprops", "hubcaps");
             let issues = repo.issues();
             for issue in issues.list(&IssueListOptions::builder().build()).unwrap() {