+++ examples/repos.rs
@@ -1,17 +1,19 @@
+extern crate env_logger;
 extern crate hyper;
 extern crate hubcaps;
 
 use hyper::Client;
-use hubcaps::Github;
+use hubcaps::{Credentials, Github};
 use std::env;
 
 fn main() {
+    env_logger::init().unwrap();
     match env::var("GITHUB_TOKEN").ok() {
         Some(token) => {
             let client = Client::new();
             let github = Github::new(format!("hubcaps/{}", env!("CARGO_PKG_VERSION")),
                                      &client,
-                                     Some(token));
+                                     Credentials::Token(token));
             for repo in github.repos().list().unwrap() {
                 println!("{:#?}", repo)
             }